requirements.txt
absl-py==1.0.0
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.0
appdirs==1.4.4
apptools==5.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blis==0.7.7
Boruta==0.3
cachetools==5.0.0
catalogue==1.0.0
catboost==1.0.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
cloudpickle==2.0.0
colorama==0.4.4
colorlover==0.3.0
configobj==5.0.6
cowsay==4.0
cryptography==36.0.2
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.6
databricks-cli==0.16.7
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dma==0.2.1
dnspython==2.2.1
docker==5.0.3
docopt==0.6.2
entrypoints==0.4
envisage==6.1.0
et-xmlfile==1.1.0
executing==0.8.3
fbm==0.3.0
Flask==2.1.0
flatbuffers==2.0
fonttools==4.30.0
frozenlist==1.3.1
funcy==1.17
future==0.18.2
gast==0.5.3
gcloud==0.18.3
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-cloud-iot==2.5.1
google-cloud-pubsub==2.13.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
graphviz==0.20
greenlet==1.1.2
grpc-google-iam-v1==0.12.4
grpcio==1.46.0
grpcio-status==1.46.0
h5py==3.6.0
htmlmin==0.1.12
httplib2==0.20.4
idna==3.3
ImageHash==4.2.1
imageio==2.22.2
imbalanced-learn==0.7.0
importlib-metadata==4.11.3
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
kmodes==0.12.1
libclang==14.0.1
lightgbm==3.3.2
llvmlite==0.37.0
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mayavi==4.8.1
missingno==0.5.1
mistune==0.8.4
mlflow==1.26.1
mlxtend==0.19.0
mne==1.1.1
multidict==6.0.2
multimethod==1.8
murmurhash==1.0.7
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.8.3
nibabel==4.0.2
nltk==3.7
notebook==6.4.10
numba==0.54.1
numexpr==2.8.1
numpy==1.23.4
numpy-stl==2.17.1
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.1
pandas-profiling==3.2.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
phik==0.12.2
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.11
plac==1.1.3
plotly==5.7.0
pooch==1.6.0
preshed==3.0.6
prometheus-client==0.13.1
prometheus-flask-exporter==0.20.2
prompt-toolkit==3.0.28
proto-plus==1.20.3
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycaret==2.3.10
pycparser==2.21
pydantic==1.9.1
pydicom==2.3.0
pyEDFlib==0.1.28
pyface==7.4.2
Pygments==2.11.2
PyJWT==2.3.0
pyLDAvis==3.2.2
pymongo==4.0.2
pynndescent==0.5.7
pyod==1.0.1
pyparsing==3.0.7
pyproject-toml==0.0.10
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyrsistent==0.18.1
pysurfer==0.11.0
python-dateutil==2.8.2
python-utils==3.4.5
pytz==2021.3
PyWavelets==1.3.0
pywin32==227
pywinpty==2.0.5
PyYAML==5.4.1
pyzmq==22.3.0
querystring-parser==1.2.4
redis==4.3.1
regex==2022.6.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
sense-emu==1.1
sense-hat==2.2.0
shap==0.40.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
smart-open==6.0.0
smmap==5.0.0
soupsieve==2.3.1
spacy==2.3.7
SQLAlchemy==1.4.37
sqlparse==0.4.2
srsly==1.0.5
stack-data==0.2.0
statsmodels==0.12.0
stl==0.0.3
tabulate==0.8.9
tangled-up-in-unicode==0.2.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
textblob==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
thinc==7.4.5
threadpoolctl==3.1.0
tifffile==2022.10.10
tk==0.1.0
toml==0.10.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
traits==6.4.1
traitsui==7.4.1
typing_extensions==4.2.0
umap-learn==0.5.3
urllib3==1.26.9
visions==0.7.4
vtk==9.2.2
waitress==2.1.2
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.0
widgetsnbextension==3.6.0
wikipedia==1.4.0
wordcloud==1.8.1
wrapt==1.14.1
wslink==1.9.0
yarg==0.1.9
yarl==1.8.1
yellowbrick==1.3.post1
zipp==3.7.0
